+++ examples/sqla-custom-inline-forms/app.py
@@ -1,7 +1,7 @@
 import os
 import os.path as op
 
-from werkzeug import secure_filename
+from werkzeug.utils import secure_filename
 from sqlalchemy import event
 
 from flask import Flask, request, render_template
@@ -19,7 +19,7 @@ from flask_admin.contrib.sqla.fields import InlineModelFormList
 # Create application
 app = Flask(__name__)
 
-# Create dummy secrey key so we can use sessions
+# Create dummy secret key so we can use sessions
 app.config['SECRET_KEY'] = '123456790'
 
 # Create in-memory database