+++ examples/sqla/admin/main.py
@@ -242,7 +242,7 @@ class TreeView(sqla.ModelView):
 
 
 # Create admin
-admin = admin.Admin(app, name='Example: SQLAlchemy', template_mode='bootstrap3')
+admin = admin.Admin(app, name='Example: SQLAlchemy', template_mode='bootstrap4')
 
 # Add views
 admin.add_view(UserAdmin(User, db.session))