+++ flask_admin/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.5.4'
+__version__ = '1.5.8'
 __author__ = 'Flask-Admin team'
 __email__ = 'serge.koval+github@gmail.com'