+++ flask_admin/model/base.py
@@ -5,11 +5,10 @@ import mimetypes
 import time
 from math import ceil
 
-from werkzeug import secure_filename
+from werkzeug.utils import secure_filename
 
 from flask import (current_app, request, redirect, flash, abort, json,
                    Response, get_flashed_messages, stream_with_context)
-from jinja2 import contextfunction
 try:
     import tablib
 except ImportError:
@@ -29,7 +28,7 @@ from flask_admin.helpers import (get_form_data, validate_form_on_submit,
 from flask_admin.tools import rec_getattr
 from flask_admin._backwards import ObsoleteAttr
 from flask_admin._compat import (iteritems, itervalues, OrderedDict,
-                                 as_unicode, csv_encode, text_type)
+                                 as_unicode, csv_encode, text_type, pass_context)
 from .helpers import prettify_name, get_mdict_item_or_list
 from .ajax import AjaxModelLoader
 
@@ -1467,7 +1466,7 @@ class BaseModelView(BaseView, ActionsMixin):
             :param name:
                 Column name.
         """
-        return name in self.column_editable_list
+        return name in self.column_editable_list and self.can_edit
 
     def _get_column_by_idx(self, idx):
         """
@@ -1851,7 +1850,7 @@ class BaseModelView(BaseView, ActionsMixin):
 
         return value
 
-    @contextfunction
+    @pass_context
     def get_list_value(self, context, model, name):
         """
             Returns the value to be displayed in the list view
@@ -1871,7 +1870,7 @@ class BaseModelView(BaseView, ActionsMixin):
             self.column_type_formatters,
         )
 
-    @contextfunction
+    @pass_context
     def get_detail_value(self, context, model, name):
         """
             Returns the value to be displayed in the detail view
@@ -2145,7 +2144,7 @@ class BaseModelView(BaseView, ActionsMixin):
                 if '_add_another' in request.form:
                     return redirect(self.get_url('.create_view', url=return_url))
                 elif '_continue_editing' in request.form:
-                    return redirect(request.url)
+                    return redirect(self.get_url('.edit_view', id=self.get_pk_value(model)))
                 else:
                     # save button
                     return redirect(self.get_save_return_url(model, is_created=False))