+++ flask_admin/model/typefmt.py
@@ -1,6 +1,6 @@
 import json
 
-from jinja2 import Markup
+from markupsafe import Markup
 from flask_admin._compat import text_type
 try:
     from enum import Enum
@@ -36,8 +36,8 @@ def bool_formatter(view, value):
             Value to check
     """
     glyph = 'ok-circle' if value else 'minus-sign'
-    fa = 'check-circle' if value else 'minus-circle'
-    return Markup('<span class="fa fa-%s glyphicon glyphicon-%s icon-%s"></span>' % (fa, glyph, glyph))
+    fa = 'fa-check-circle' if value else 'fa-minus-circle'
+    return Markup('<span class="fa %s glyphicon glyphicon-%s icon-%s"></span>' % (fa, glyph, glyph))
 
 
 def list_formatter(view, values):